+++ packages/vite/src/node/server/transformRequest.ts
@@ -5,6 +5,7 @@ import getEtag from 'etag'
 import convertSourceMap from 'convert-source-map'
 import type { SourceDescription, SourceMap } from 'rollup'
 import colors from 'picocolors'
+import MagicString from 'magic-string'
 import type { ViteDevServer } from '..'
 import {
   blankReplacer,
@@ -18,8 +19,13 @@ import {
 } from '../utils'
 import { checkPublicFile } from '../plugins/asset'
 import { getDepsOptimizer } from '../optimizer'
+import { isCSSRequest } from '../plugins/css'
+import { SPECIAL_QUERY_RE } from '../constants'
 import { injectSourcesContent } from './sourcemap'
 import { isFileServingAllowed } from './middlewares/static'
+
+export const ERR_LOAD_URL = 'ERR_LOAD_URL'
+export const ERR_LOAD_PUBLIC_URL = 'ERR_LOAD_PUBLIC_URL'
 
 const debugLoad = createDebugger('vite:load')
 const debugTransform = createDebugger('vite:transform')
@@ -215,18 +221,18 @@ async function loadAndTransform(
     }
   }
   if (code == null) {
-    if (checkPublicFile(url, config)) {
-      throw new Error(
-        `Failed to load url ${url} (resolved id: ${id}). ` +
-          `This file is in /public and will be copied as-is during build without ` +
-          `going through the plugin transforms, and therefore should not be ` +
-          `imported from source code. It can only be referenced via HTML tags.`
-      )
-    } else {
-      return null
-    }
+    const isPublicFile = checkPublicFile(url, config)
+    const msg = isPublicFile
+      ? `This file is in /public and will be copied as-is during build without ` +
+        `going through the plugin transforms, and therefore should not be ` +
+        `imported from source code. It can only be referenced via HTML tags.`
+      : `Does the file exist?`
+    const err: any = new Error(
+      `Failed to load url ${url} (resolved id: ${id}). ${msg}`
+    )
+    err.code = isPublicFile ? ERR_LOAD_PUBLIC_URL : ERR_LOAD_URL
+    throw err
   }
-
   // ensure module in graph after successful load
   const mod = await moduleGraph.ensureEntryFromUrl(url, ssr)
   ensureWatchedFile(watcher, mod.file, root)
@@ -251,11 +257,26 @@ async function loadAndTransform(
     isDebug && debugTransform(`${timeFrom(transformStart)} ${prettyUrl}`)
     code = transformResult.code!
     map = transformResult.map
+
+    // To enable IDE debugging, add a minimal sourcemap for modified JS files without one
+    if (
+      !map &&
+      mod.file &&
+      mod.type === 'js' &&
+      code !== originalCode &&
+      !(isCSSRequest(id) && !SPECIAL_QUERY_RE.test(id)) // skip CSS : #9914
+    ) {
+      map = new MagicString(code).generateMap({ source: mod.file })
+    }
   }
 
   if (map && mod.file) {
     map = (typeof map === 'string' ? JSON.parse(map) : map) as SourceMap
-    if (map.mappings && !map.sourcesContent) {
+    if (
+      map.mappings &&
+      (!map.sourcesContent ||
+        (map.sourcesContent as Array<string | null>).includes(null))
+    ) {
       await injectSourcesContent(map, mod.file, logger)
     }
   }